refactor(site,demo): Site & component demo integration
* Move component demos into src/__demo__, thus including them in the package/dist
* Update webpack to build using new demo locations
* Render Hello demo basic example in site
* Add build:watch npm script to enable auto-reloading in site/demos
