+++ package.json
@@ -7,6 +7,7 @@
     "bootstrap": "lerna bootstrap && npm run link-parent-bin",
     "build": "lerna run build",
     "build:prod": "lerna run build:prod",
+    "build:watch": "onchange 'packages/**/*.js' -e '**/dist/**' -i -- sh -c 'npm run bootstrap && npm run build'",
     "clean": "lerna clean --yes && lerna run clean",
     "commit": "git-cz",
     "commitmsg": "validate-commit-msg",
@@ -46,6 +47,7 @@
     "jest": "19.0.2",
     "lerna": "2.0.0-rc.3",
     "link-parent-bin": "0.1.1",
+    "onchange": "3.2.1",
     "prettier": "1.0.2",
     "react": "15.4.2",
     "react-addons-test-utils": "15.4.2",
