+++ packages/components/archetype/src/__demo__/examples/Basic.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Archetype from '../../src/Archetype';
+import Archetype from '../../Archetype';
 
 export default function Basic() {
   return <Archetype />;
