+++ packages/components/hello-world/src/__demo__/App.js
