+++ packages/components/hello/src/__demo__/examples/Basic.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Hello from '../../src/Hello';
+import Hello from '../../Hello';
 
 export default function Basic() {
   return <Hello />;
