+++ packages/components/world/src/__demo__/examples/Basic.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import World from '../../src/World';
+import World from '../../World';
 
 export default function Basic() {
   return <World />;
