+++ packages/site/src/App.js
@@ -1,9 +1,12 @@
 import React from 'react';
+import BasicHelloExample
+  from '@mineral-ui/hello/dist/es/__demo__/examples/Basic';
 
 export default function App() {
   return (
     <div>
-      Site
+      <h1>Site</h1>
+      <BasicHelloExample />
     </div>
   );
 }
